requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
cytoolz==0.8.2
Django==1.11.6
django-anymail==1.4
django-appconf==1.0.2
django-ckeditor==5.3.1
django-colorful==1.2
django-cors-headers==2.2.0
django-countries==5.0
django-embed-video==1.1.2
django-formtools==2.1
django-js-asset==0.1.1
django-modeltranslation==0.12.1
django-nested-admin==3.0.21
django-otp==0.4.1.1
django-phonenumber-field==1.3.0
django-rosetta==0.7.13
django-solo==1.1.2
django-two-factor-auth==1.6.2
docopt==0.4.0
ethereum-abi-utils==0.4.4
ethereum-keys==0.1.0a7
ethereum-tester==0.1.0b2
ethereum-utils==0.6.0
idna==2.6
mailchimp==2.0.9
microsofttranslator==0.8
MySQL-python==1.2.5
olefile==0.44
phonenumberslite==8.8.6
Pillow==4.3.0
pinax-referrals==3.0.3
pkg-resources==0.0.0
polib==1.0.8
pysha3==1.0.2
python-monkey-business==1.0.0
pytz==2017.3
pyuca==1.2
qrcode==4.0.4
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
sorl-thumbnail==12.3
tinify==1.5.1
toolz==0.8.2
urllib3==1.22
